Format solidity

+++ packages/arb-bridge-eth/contracts/GlobalInbox.sol
@@ -31,7 +31,13 @@ import "./arch/Value.sol";
 import "./libraries/SigUtils.sol";
 import "./libraries/BytesLib.sol";
 
-contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGlobalInbox, PaymentRecords {
+contract GlobalInbox is
+    GlobalEthWallet,
+    GlobalFTWallet,
+    GlobalNFTWallet,
+    IGlobalInbox,
+    PaymentRecords
+{
     uint8 internal constant TRANSACTION_MSG = 0;
     uint8 internal constant ETH_DEPOSIT = 1;
     uint8 internal constant ERC20_DEPOSIT = 2;
@@ -50,12 +56,20 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
 
     mapping(address => Inbox) private inboxes;
 
-    function getInbox(address account) external view returns(bytes32, uint) {
+    function getInbox(address account)
+        external
+        view
+        returns (bytes32, uint256)
+    {
         Inbox storage inbox = inboxes[account];
         return (inbox.value, inbox.count);
     }
 
-    function sendMessages(bytes calldata _messages, uint[] calldata messageCounts, bytes32[] calldata nodeHashes) external {
+    function sendMessages(
+        bytes calldata _messages,
+        uint256[] calldata messageCounts,
+        bytes32[] calldata nodeHashes
+    ) external {
         bool valid;
         uint256 offset = 0;
         uint256 messageType;
@@ -64,15 +78,18 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         uint256 nodeCount = nodeHashes.length;
         for (uint256 i = 0; i < nodeCount; i++) {
             for (uint256 j = 0; j < messageCounts[i]; j++) {
-                (   valid,
-                    offset,
-                    messageType,
-                    sender
-                ) = Value.deserializeMessageData(_messages, offset);
+                (valid, offset, messageType, sender) = Value
+                    .deserializeMessageData(_messages, offset);
                 if (!valid) {
                     return;
                 }
-                (valid, offset) = sendDeserializedMsg(nodeHashes[i], j, _messages, offset, messageType);
+                (valid, offset) = sendDeserializedMsg(
+                    nodeHashes[i],
+                    j,
+                    _messages,
+                    offset,
+                    messageType
+                );
                 if (!valid) {
                     return;
                 }
@@ -88,18 +105,14 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         uint256 messageType
     )
         private
-        returns(
+        returns (
             bool, // valid
             uint256 // offset
         )
     {
         if (messageType == ETH_DEPOSIT) {
-            (
-                bool valid,
-                uint256 offset,
-                address to,
-                uint256 value
-            ) = Value.getEthMsgData(_messages, startOffset);
+            (bool valid, uint256 offset, address to, uint256 value) = Value
+                .getEthMsgData(_messages, startOffset);
 
             if (!valid) {
                 return (false, startOffset);
@@ -155,9 +168,7 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         uint256 _seqNumber,
         uint256 _value,
         bytes calldata _data
-    )
-        external
-    {
+    ) external {
         _deliverTransactionMessage(
             _chain,
             _to,
@@ -171,12 +182,7 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
     function depositEthMessage(address _chain, address _to) external payable {
         depositEth(_chain);
 
-        _deliverEthMessage(
-            _chain,
-            _to,
-            msg.sender,
-            msg.value
-        );
+        _deliverEthMessage(_chain, _to, msg.sender, msg.value);
     }
 
     function depositERC20Message(
@@ -184,18 +190,10 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         address _to,
         address _erc20,
         uint256 _value
-    )
-        external
-    {
+    ) external {
         depositERC20(_erc20, _chain, _value);
 
-        _deliverERC20TokenMessage(
-            _chain,
-            _to,
-            msg.sender,
-            _erc20,
-            _value
-        );
+        _deliverERC20TokenMessage(_chain, _to, msg.sender, _erc20, _value);
     }
 
     function depositERC721Message(
@@ -203,18 +201,10 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         address _to,
         address _erc721,
         uint256 _id
-    )
-        external
-    {
+    ) external {
         depositERC721(_erc721, _chain, _id);
 
-        _deliverERC721TokenMessage(
-            _chain,
-            _to,
-            msg.sender,
-            _erc721,
-            _id
-        );
+        _deliverERC721TokenMessage(_chain, _to, msg.sender, _erc721, _id);
     }
 
     function forwardContractTransactionMessage(
@@ -222,9 +212,7 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         address _from,
         uint256 _value,
         bytes calldata _data
-    )
-        external
-    {
+    ) external {
         _deliverContractTransactionMessage(
             msg.sender,
             _to,
@@ -237,12 +225,7 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
     function forwardEthMessage(address _to, address _from) external payable {
         depositEth(msg.sender);
 
-        _deliverEthMessage(
-            msg.sender,
-            _to,
-            _from,
-            msg.value
-        );
+        _deliverEthMessage(msg.sender, _to, _from, msg.value);
     }
 
     // // Transaction format
@@ -253,19 +236,12 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
     // //   signature (65 bytes)
     // //   data (arbitrary length)
 
-
-    function deliverTransactionBatch(
-        address chain,
-        bytes calldata transactions
-    )
+    function deliverTransactionBatch(address chain, bytes calldata transactions)
         external
     {
         require(msg.sender == tx.origin, "origin only");
         bytes32 messageHash = keccak256(
-            abi.encodePacked(
-                TRANSACTION_BATCH_MSG,
-                transactions
-            )
+            abi.encodePacked(TRANSACTION_BATCH_MSG, transactions)
         );
 
         _deliverMessage(chain, messageHash);
@@ -280,9 +256,7 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         uint256 _seqNumber,
         uint256 _value,
         bytes memory _data
-    )
-        private
-    {
+    ) private {
         bytes32 messageHash = Messages.transactionHash(
             _chain,
             _to,
@@ -309,14 +283,8 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         address _to,
         address _from,
         uint256 _value
-    )
-        private
-    {
-        bytes32 messageHash = Messages.ethHash(
-            _to,
-            _from,
-            _value
-        );
+    ) private {
+        bytes32 messageHash = Messages.ethHash(_to, _from, _value);
 
         uint256 messageNum = _deliverMessage(_chain, messageHash);
 
@@ -335,15 +303,8 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         address _from,
         address _erc20,
         uint256 _value
-    )
-        private
-    {
-        bytes32 messageHash = Messages.erc20Hash(
-            _to,
-            _from,
-            _erc20,
-            _value
-        );
+    ) private {
+        bytes32 messageHash = Messages.erc20Hash(_to, _from, _erc20, _value);
 
         uint256 messageNum = _deliverMessage(_chain, messageHash);
 
@@ -363,15 +324,8 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         address _from,
         address _erc721,
         uint256 _id
-    )
-        private
-    {
-        bytes32 messageHash = Messages.erc721Hash(
-            _to,
-            _from,
-            _erc721,
-            _id
-        );
+    ) private {
+        bytes32 messageHash = Messages.erc721Hash(_to, _from, _erc721, _id);
 
         uint256 messageNum = _deliverMessage(_chain, messageHash);
 
@@ -391,9 +345,7 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         address _from,
         uint256 _value,
         bytes memory _data
-    )
-        private
-    {
+    ) private {
         bytes32 messageHash = Messages.contractTransactionHash(
             _to,
             _from,
@@ -413,7 +365,10 @@ contract GlobalInbox is GlobalEthWallet, GlobalFTWallet, GlobalNFTWallet, IGloba
         );
     }
 
-    function _deliverMessage(address _chain, bytes32 _messageHash) private returns(uint256) {
+    function _deliverMessage(address _chain, bytes32 _messageHash)
+        private
+        returns (uint256)
+    {
         Inbox storage inbox = inboxes[_chain];
         uint256 updatedCount = inbox.count + 1;
         inbox.value = Messages.addMessageToInbox(
