+++ packages/arb-bridge-eth/contracts/arch/Machine.sol
@@ -22,7 +22,6 @@ import "./Value.sol";
 
 import "../libraries/DebugPrint.sol";
 
-
 library Machine {
     using Value for Value.Data;
 
@@ -33,11 +32,7 @@ library Machine {
     function addStackVal(
         Value.Data memory stackValHash,
         Value.Data memory valHash
-    )
-        internal
-        pure
-        returns (Value.Data memory)
-    {
+    ) internal pure returns (Value.Data memory) {
         Value.Data[] memory vals = new Value.Data[](2);
         vals[0] = valHash;
         vals[1] = stackValHash;
@@ -87,19 +82,22 @@ library Machine {
         machine.status = MACHINE_HALT;
     }
 
-    function addDataStackValue(Data memory machine, Value.Data memory val) internal pure {
+    function addDataStackValue(Data memory machine, Value.Data memory val)
+        internal
+        pure
+    {
         machine.dataStack = addStackVal(machine.dataStack, val);
     }
 
-    function addAuxStackValue(Data memory machine, Value.Data memory val) internal pure {
+    function addAuxStackValue(Data memory machine, Value.Data memory val)
+        internal
+        pure
+    {
         machine.auxStack = addStackVal(machine.auxStack, val);
     }
 
     function addDataStackInt(Data memory machine, uint256 val) internal pure {
-        machine.dataStack = addStackVal(
-            machine.dataStack,
-            Value.newInt(val)
-        );
+        machine.dataStack = addStackVal(machine.dataStack, Value.newInt(val));
     }
 
     function machineHash(
@@ -129,9 +127,9 @@ library Machine {
 
     function hash(Data memory machine) internal pure returns (bytes32) {
         if (machine.status == MACHINE_HALT) {
-            return bytes32(uint(0));
+            return bytes32(uint256(0));
         } else if (machine.status == MACHINE_ERRORSTOP) {
-            return bytes32(uint(1));
+            return bytes32(uint256(1));
         } else {
             return keccak256(
                 abi.encodePacked(
@@ -144,7 +142,6 @@ library Machine {
                 )
             );
         }
-
     }
 
     function clone(Data memory machine) internal pure returns (Data memory) {
@@ -159,13 +156,10 @@ library Machine {
         );
     }
 
-    function deserializeMachine(
-        bytes memory data,
-        uint256 offset
-    )
+    function deserializeMachine(bytes memory data, uint256 offset)
         internal
         pure
-        returns(
+        returns (
             bool, // valid
             uint256, // offset
             Data memory // machine
@@ -179,11 +173,17 @@ library Machine {
             return (false, offset, m);
         }
 
-        (valid, offset, m.dataStack) = Value.deserializeHashPreImage(data, offset);
+        (valid, offset, m.dataStack) = Value.deserializeHashPreImage(
+            data,
+            offset
+        );
         if (!valid) {
             return (false, offset, m);
         }
-        (valid, offset, m.auxStack) = Value.deserializeHashPreImage(data, offset);
+        (valid, offset, m.auxStack) = Value.deserializeHashPreImage(
+            data,
+            offset
+        );
         if (!valid) {
             return (false, offset, m);
         }
