+++ packages/arb-bridge-eth/contracts/vm/VM.sol
@@ -21,14 +21,14 @@ pragma solidity ^0.5.11;
 import "../arch/Value.sol";
 import "../libraries/SafeMath.sol";
 
-
 library VM {
     using SafeMath for uint256;
 
     bytes32 private constant MACHINE_HALT_HASH = bytes32(0);
-    bytes32 private constant MACHINE_ERROR_HASH = bytes32(uint(1));
+    bytes32 private constant MACHINE_ERROR_HASH = bytes32(uint256(1));
 
-    struct Params {  // these are defined just once for each vM
+    struct Params {
+        // these are defined just once for each vM
         uint256 gracePeriodTicks;
         uint256 arbGasSpeedLimitPerTick;
         uint64 maxExecutionSteps;
@@ -36,15 +36,19 @@ library VM {
         uint64 maxTimestampBoundsWidth;
     }
 
-    function isErrored(bytes32 vmStateHash) internal pure returns(bool) {
+    function isErrored(bytes32 vmStateHash) internal pure returns (bool) {
         return vmStateHash == MACHINE_ERROR_HASH;
     }
 
-    function isHalted(bytes32 vmStateHash) internal pure returns(bool) {
+    function isHalted(bytes32 vmStateHash) internal pure returns (bool) {
         return vmStateHash == MACHINE_HALT_HASH;
     }
 
-    function withinTimeBounds(uint128[4] memory _timeBoundsBlocks) internal view returns (bool) {
+    function withinTimeBounds(uint128[4] memory _timeBoundsBlocks)
+        internal
+        view
+        returns (bool)
+    {
         return
             block.number >= _timeBoundsBlocks[0] &&
             block.number <= _timeBoundsBlocks[1] &&
